book-recommendation-system/requirements.txt
pandas
numpy
scikit-learn
surprise
fastapi
uvicorn[standard]
python-dotenv
jupyter